refactor: wallpaper 改用 LibGDX 官方 AndroidLiveWallpaperService；国内镜像；显式 buildToolsVersion；Color 透明度修正

+++ app/src/main/java/com/zhuomo/flowlume/app/ui/screens/RenderScreen.kt
@@ -112,7 +112,7 @@ fun RenderScreen() {
                         modifier = Modifier
                             .size(30.dp)
                             .clip(RoundedCornerShape(8.dp))
-                            .background(Color(tone.toInt()))
+                            .background(Color(tone))
                             .clickable { update(config.copy(defaultTone = tone)) }
                             .padding(2.dp)
                     ) {
